Inits the new version of portfolio
I decided to start over. However, I needed to
maintain a quick, handy, and easily accessible
copy of the old portfolio. Decided to move the
old portfolio to the old directory and the new
version is on the new directory.

Will eventually delete the old directory and move
everything in the new directory out to the root
of the project.