Styling fixes
Makes max-width 980px as that seems
to be optimized for Apple. The Apple
site is based on content max width of
980px.

Also, the background images were really
zoomed in and the parallax effect was not
great on the iPhone15 Pro Max when holding
it horizontally. This commit hopes to fix
that by setting the background-attachment
of images to scroll when the max-width is
950px. This works on the Chrome browser
simulator but from experience this is not
a very accurate simulator.